move project table
